Removed duplicate script.
Watch-all-event-contents.py was doing more or less the same as watch-all-
events.py, the difference being it was printing out more detail. I removed the
former and moved the extra bits to the latter (which we have been shipping so
far).
